ICommentRepository was split in ICommentRepository and IPostRepository.
