+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/app/AppModule.kt
@@ -9,7 +9,8 @@ import org.koin.dsl.module
 val appModule = module {
     single { provideRetrofit() }
     single <ICommentSourceData> { provideNetworkApi( get() ) }
+    single <IPostRepository> { PostRepository( get() ) }
     single <ICommentRepository> { CommentRepository( get() ) }
-    factory <MiddleCommentsUserCase> { MiddleCommentsUserCase( get() ) }
+    factory <MiddleCommentsUserCase> { MiddleCommentsUserCase( get(), get() ) }
     viewModel { MiddleCommentViewModel( get() ) }
 }
