+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/CommentRepository.kt
@@ -1,16 +1,8 @@
 package ar.scacchipa.jsonplaceholderviewer.data
 
 class CommentRepository(
-     var apiSourceData: ICommentSourceData// = provideNetworkApi(provideRetrofit())
-): ICommentRepository {
-    override suspend fun getPosts(): List<Post> {
-        val response = apiSourceData.getPosts()
-        return if (response.isSuccessful) {
-            response.body() ?: listOf()
-        } else {
-            listOf()
-        }
-    }
+    var apiSourceData: ICommentSourceData// = provideNetworkApi(provideRetrofit())
+): ICommentRepository  {
     override suspend fun getComment(id: Int): List<Comment> {
         apiSourceData.getComment(id.toString()).let { response ->
             return if (response.isSuccessful) {
