+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/ICommentRepository.kt
@@ -1,6 +1,5 @@
 package ar.scacchipa.jsonplaceholderviewer.data
 
 interface ICommentRepository {
-    suspend fun getPosts(): List<Post>
     suspend fun getComment(id:Int): List<Comment>
 }
