+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/IPostRepository.kt
@@ -1,0 +1,5 @@
+package ar.scacchipa.jsonplaceholderviewer.data
+
+interface IPostRepository {
+    suspend fun getPosts(): List<Post>
+}
